+++ src/main/java/com/example/haveyoursay/services/PetitionService.java
@@ -3,5 +3,5 @@ package com.example.haveyoursay.services;
 import com.example.haveyoursay.models.Petition;
 
 public interface PetitionService {
-    public Integer getPetitionCloseTime(Integer petitionStartTime);
+    public Long getPetitionCloseTime(Long petitionStartTime);
 }
